Adding Classes, price contents, function adaptations
In this session, changes were made to the color variables in the general.css.

To the admin dashboards, earnings were added to it

To the admissionForm.js, its processes were connected directly to the database for communication with live server

Changes were made to the loadDisplay function. Its parameters were changed to an array to help specify a needed part of the loader

Files like password.php have being made to adapt to the change in the loader function

changes to the transaction table and also the cssps table has been made to the submit.php file